+++ rust-spice/src/core/das.rs
@@ -1,4 +1,0 @@
-/*!
-# SPICE functions for the DAS subsystem
-
-*/